🚀 Deploy ScuffBot main to Kube
This PR sets up ScuffBot to be ready to be deployed to Kube.

This involves creating the necessary workflows. These workflows will
have the bot automatically be deployed to separate dev and prod envs for
testing, with the dev environment needing to be successfully deployed to
for the PR check to pass.

Other changes include removal of secrets from the `config.yaml`, and
instead relying on a dedicated secrets folder, as well as removing
redundant imports through the repo.

The repo also now makes use of using `docker-compose` to run locally,
for ease of setup and future changes.