💚 Consolidate build and deploy pipeline
Consolidating build and deploy pipeline into a single pipeline to avoid
issues with trying to figure out the correct namespace to deploy to
(with the old workflow_run approach), as well as have the entire process
be reported in the branch's CI checks.

Additionally, an option has been added to `workflow_dispatch` to choose
whether to deploy to `staging` or `prod`, say for feature branch deployments
in the future.

'Dev' has also been renamed to 'staging' to better reflect the
environment's purpose.